Refactoring and bug fixes parser
Replace Interop.Word by Spire.Doc
Bug fix get number days from parser
Refactoring ChudoPechka parser
