+++ Food.Api/Web.config
@@ -13,7 +13,7 @@
   </connectionStrings>
   <appSettings file="Private.config"></appSettings>
   <system.web>
-    <customErrors mode="Off"/>
+    <customErrors mode="Off" />
     <authentication mode="None" />
     <compilation debug="true" targetFramework="4.5.2" />
     <httpRuntime targetFramework="4.5.2" />
