+++ Food.Api/packages.config
@@ -6,6 +6,7 @@
   <package id="EnterpriseLibrary.TransientFaultHandling.Data" version="6.0.1304.1" targetFramework="net452" />
   <package id="EntityFramework" version="6.1.3" targetFramework="net452" />
   <package id="EntityFramework.ru" version="6.1.3" targetFramework="net452" />
+  <package id="FreeSpire.Doc" version="5.5" targetFramework="net452" />
   <package id="HtmlAgilityPack-PCL" version="1.4.6" targetFramework="net452" />
   <package id="jQuery" version="1.10.2" targetFramework="net452" />
   <package id="jQuery.Validation" version="1.11.1" targetFramework="net452" />
